Read GAMS solver and model status
Determine the most recent gdx from the main.lst and read model and solver status from this gdx.
In standard brick run: Stop if there was a run error or the model is infeasible.
In calibration run: If step size adaptation is infeasible, continue but make sure the respective step size is rejected. If any other run during the calibration is infeasible, stop but still write diagnostic output.